corrigiendoTablas
