+++ src/main/java/pe/edu/upc/aaw/demoproyecto/dtos/PagoDTO.java
@@ -1,5 +1,9 @@
 package pe.edu.upc.aaw.demoproyecto.dtos;
 
+import pe.edu.upc.aaw.demoproyecto.entities.Card;
+import pe.edu.upc.aaw.demoproyecto.entities.Membresia;
+
+import javax.persistence.*;
 import java.time.LocalDate;
 
 public class PagoDTO {
@@ -7,6 +11,16 @@ public class PagoDTO {
     private int quantityPago;
     private boolean payConfirmed;
     private LocalDate datePago;
+    private Membresia membresia;
+    private Card card;
+
+    public Card getCard() {
+        return card;
+    }
+
+    public void setCard(Card card) {
+        this.card = card;
+    }
 
     public int getId() {
         return id;
@@ -38,5 +52,13 @@ public class PagoDTO {
 
     public void setDatePago(LocalDate datePago) {
         this.datePago = datePago;
+    }
+
+    public Membresia getMembresia() {
+        return membresia;
+    }
+
+    public void setMembresia(Membresia membresia) {
+        this.membresia = membresia;
     }
 }
