+++ src/main/java/pe/edu/upc/aaw/demoproyecto/entities/Pago.java
@@ -17,14 +17,26 @@ public class Pago {
     @Column(name = "datePago",nullable = false)
     private LocalDate datePago;
 
-    public Pago(int id, int quantityPago, boolean payConfirmed, LocalDate datePago) {
+    @ManyToOne
+    @JoinColumn(name = "idMembresia")
+    private Membresia membresia;
+
+    @OneToOne
+    @JoinColumn(name = "idCard")
+    private Card card;
+
+    public Pago() {
+    }
+
+
+    public Pago(int id, int quantityPago, boolean payConfirmed, LocalDate datePago, Membresia membresia, Card card) {
         this.id = id;
         this.quantityPago = quantityPago;
         this.payConfirmed = payConfirmed;
         this.datePago = datePago;
+        this.membresia = membresia;
+        this.card = card;
     }
-
-    public Pago() {}
 
     public int getId() {
         return id;
@@ -56,5 +68,21 @@ public class Pago {
 
     public void setDatePago(LocalDate datePago) {
         this.datePago = datePago;
+    }
+
+    public Membresia getMembresia() {
+        return membresia;
+    }
+
+    public void setMembresia(Membresia membresia) {
+        this.membresia = membresia;
+    }
+
+    public Card getCard() {
+        return card;
+    }
+
+    public void setCard(Card card) {
+        this.card = card;
     }
 }
