+++ src/main/resources/application.properties
@@ -4,7 +4,6 @@ spring.jpa.hibernate.ddl-auto=update
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.datasource.url=jdbc:postgresql://localhost/StreamGO
 spring.datasource.username=postgres
-spring.datasource.password=kurtyvania
 spring.datasource.password=admin
 
 spring.mvc.pathmatch.matching-strategy = ANT_PATH_MATCHER
